+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Constants/PluginSetupDataEntrySlugs.php
@@ -14,7 +14,7 @@ class PluginSetupDataEntrySlugs
      * @gatographql-note: Do not rename this slug, as it's referenced when installing the testing webservers
      */
     public final const SCHEMA_CONFIGURATION_NESTED_MUTATIONS_ENTITY_AS_MUTATION_PAYLOAD_TYPE = 'nested-mutations-entity-as-mutation-payload-type';
-    
+
     public final const CUSTOM_ENDPOINT_NESTED_MUTATIONS = 'nested-mutations';
     public final const CUSTOM_ENDPOINT_NESTED_MUTATIONS_ENTITY_AS_MUTATION_PAYLOAD_TYPE = 'nested-mutations-entity-as-mutation-payload-type';