+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/AbstractPlugin.php
@@ -370,10 +370,17 @@ abstract class AbstractPlugin implements PluginInterface
     }
 
     /**
-     * Execute logic after the plugin/extension has just been activated
-     * (for first time)
+     * Execute logic after the plugin/extension has just been activated.
+     *
+     * Notice that this will be executed when first time activated, or
+     * reactivated (i.e. activated => deactivated => activated).
+     *
+     * Then, when installing setup data, we must first check that the entry
+     * does not already exist. This will also avoid duplicating setup data
+     * when downgrading the plugin to a lower version, and then upgrading
+     * again.
      */
-    public function pluginJustFirstTimeActivated(): void
+    public function pluginJustActivated(): void
     {
         /**
          * Taxonomies are registered on "init", hence must insert
@@ -462,7 +469,7 @@ abstract class AbstractPlugin implements PluginInterface
      *
      * The plugin's setup data will be installed if:
      *
-     * - $previousVersion = null => Activating the plugin for first time
+     * - $previousVersion = null => Activating the plugin
      * - $previousVersion < someVersion => Updating to a new version that has data to install
      */
     protected function installPluginSetupData(?string $previousVersion = null): void